codebase/backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=4.5.0
pydantic>=2.0.0
python-multipart>=0.0.5
python-dotenv>=1.0.0
requests>=2.28.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.0
bcrypt<4.0.0